# ARROW-10591: [Rust] Improve take (-20 to -40%)