--- a/rust/arrow/src/array/transform/list.rs
+++ b/rust/arrow/src/array/transform/list.rs
@@ -73,3 +73,16 @@ pub(super) fn build_extend<T: OffsetSizeTrait>(array: &ArrayData) -> Extend {
         )
     }
 }
+
+pub(super) fn extend_nulls<T: OffsetSizeTrait>(
+    mutable: &mut _MutableArrayData,
+    len: usize,
+) {
+    let mutable_offsets = mutable.buffer::<T>(0);
+    let last_offset = mutable_offsets[mutable_offsets.len() - 1];
+
+    let offset_buffer = &mut mutable.buffers[0];
+
+    let offsets = vec![last_offset; len];
+    offset_buffer.extend_from_slice(offsets.to_byte_slice());
+}