--- a/rust/arrow/src/array/transform/primitive.rs
+++ b/rust/arrow/src/array/transform/primitive.rs
@@ -33,3 +33,12 @@ pub(super) fn build_extend<T: ArrowNativeType>(array: &ArrayData) -> Extend {
         },
     )
 }
+
+pub(super) fn extend_nulls<T: ArrowNativeType>(
+    mutable: &mut _MutableArrayData,
+    len: usize,
+) {
+    let buffer = &mut mutable.buffers[0];
+    let bytes = vec![0u8; len * size_of::<T>()];
+    buffer.extend_from_slice(&bytes);
+}