… rust/arrow/src/array/transform/variable_size.rs
… rust/arrow/src/buffer.rs
… rust/arrow/src/compute/kernels/filter.rs
… rust/arrow/src/compute/kernels/take.rs
… rust/arrow/src/compute/util.rs