@drusso 2020-11-15
This looks like a neat take on take (pardon the pun!). Is there anything in particular preventing adoption of this approach?