@jorgecarleitao 2020-11-16
Performance: with the fix on the benchmarks for `take`, this is 2x slower than master.